api/requirements.txt
fastapi[standard]==0.115.*
motor==3.6.*
pymongo[srv]>=4.9,<4.10
python-dotenv==1.1.*
dnspython==2.7.*
